+++ lib/fileUtility.php
@@ -20,7 +20,7 @@
 require_once __DIR__ . "/../../../main/inc/global.inc.php";
 
 class FileUtility {
- 
+
     /**
      * Application name
      */
@@ -32,9 +32,9 @@ class FileUtility {
      * @var array
      */
     public static $can_edit_types = [
-        "docx", 
-        "xlsx", 
-        "pptx", 
+        "docx",
+        "xlsx",
+        "pptx",
         "ppsx"
     ];
 
@@ -80,12 +80,9 @@ class FileUtility {
 
     /**
      * Return file type by extension
-     *
-     * @param string $extension - extension of file
-     *
-     * @return string
      */
-    public static function getDocType($extension) {
+    public static function getDocType(string $extension): string
+    {
         if (in_array($extension, self::$text_doc)) {
             return "text";
         }
@@ -101,12 +98,9 @@ class FileUtility {
 
     /**
      * Return file extension by file type
-     *
-     * @param string $type - type of file
-     *
-     * @return string
      */
-    public static function getDocExt($type) {
+    public static function getDocExt(string $type): string
+    {
         if ($type === "text") {
             return "docx";
         }
@@ -122,16 +116,9 @@ class FileUtility {
 
     /**
      * Return file url for download
-     *
-     * @param int $courseId - identifier of course
-     * @param int $userId - identifier of user
-     * @param int $docId - identifier of document
-     * @param int $sessionId - identifier of session
-     * @param int $groupId - identifier of group or null if file out of group
-     *
-     * @return string
      */
-    public static function getFileUrl($courseId, $userId, $docId, $sessionId, $groupId) {
+    public static function getFileUrl(int $courseId, int $userId, int $docId, int $sessionId = null, int $groupId = null): string
+    {
 
         $data = [
             "type" => "download",
@@ -147,23 +134,14 @@ class FileUtility {
 
         $hashUrl = Crypt::GetHash($data);
 
-        $url = api_get_path(WEB_PLUGIN_PATH) . self::app_name . "/" . "callback.php?hash=" . $hashUrl;
-
-        return $url;
+        return api_get_path(WEB_PLUGIN_PATH) . self::app_name . "/" . "callback.php?hash=" . $hashUrl;
     }
 
     /**
      * Return file key
-     *
-     * @param string $courseCode - identifier of course
-     * @param int $userId - identifier of user
-     * @param int $docId - identifier of document
-     * @param int $sessionId - identifier of session
-     * @param int $groupId - identifier of group or null if file out of group
-     *
-     * @return string
      */
-    public static function getKey($courseCode, $docId) {
+    public static function getKey(string $courseCode, int $docId): string
+    {
         $docInfo = DocumentManager::get_document_data_by_id($docId, $courseCode);
         $mtime = filemtime($docInfo["absolute_path"]);
 
@@ -173,14 +151,11 @@ class FileUtility {
 
     /**
      * Translation key to a supported form
-     *
-     * @param string $expected_key - Expected key
-     *
-     * @return string
      */
-    public static function GenerateRevisionId($expected_key) {
-        if (strlen($expected_key) > 20) $expected_key = crc32( $expected_key);
-        $key = preg_replace("[^0-9-.a-zA-Z_=]", "_", $expected_key);
+    public static function GenerateRevisionId(string $expectedKey): string
+    {
+        if (strlen($expectedKey) > 20) $expectedKey = crc32( $expectedKey);
+        $key = preg_replace("[^0-9-.a-zA-Z_=]", "_", $expectedKey);
         $key = substr($key, 0, min(array(strlen($key), 20)));
         return $key;
     }