+++ lib/langManager.php
@@ -23,10 +23,9 @@ class LangManager {
 
     /**
      * Return lang info for current user
-     *
-     * @return array
      */
-    public static function getLangUser() {
+    public static function getLangUser(): array
+    {
         $langInfo = [];
         $userLang = api_get_language_from_type("user_profil_lang");
         if (empty($userLang)) {