+++ lib/onlyofficeActionObserver.php
@@ -22,7 +22,8 @@ class OnlyofficeActionObserver extends HookObserver implements HookDocumentActio
     /**
      * Constructor.
      */
-    public function __construct() {
+    public function __construct()
+    {
         parent::__construct(
             "plugin/onlyoffice/lib/onlyofficePlugin.php",
             "onlyoffice"
@@ -34,7 +35,8 @@ class OnlyofficeActionObserver extends HookObserver implements HookDocumentActio
      *
      * @param HookDocumentActionEventInterface $event - the hook event
      */
-    public function notifyDocumentAction(HookDocumentActionEventInterface $event) {
+    public function notifyDocumentAction(HookDocumentActionEventInterface $event)
+    {
         $data = $event->getEventData();
 
         if ($data["type"] === HOOK_EVENT_TYPE_PRE) {