+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Change Log
 
-## Added
+This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoffice-chamilo.
+
+## 2021-08 1.1.1
+- Add security filtering
+- Minor documentation and code style changes
+
+## 2021-04 1.1.0
 - View option for DOCX, XLSX, PPTX.
 - JWT support
 
-## 1.0.0
-## Added
+## 2021-03 1.0.0
 - Ability to create documents
 - Edit option for DOCX, XLSX, PPTX.
 - Collaboration editing